+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to Pipelines will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.0.18] - 2018-08-07
 ### Added
 - Add `--docker-zap` flag kill and clean all pipeline docker
   containers at once